+++ docs_src/configurations/template/example1.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 from esmerald import Esmerald
-from esmerald.config.template import TemplateConfig
+from esmerald.core.config.template import TemplateConfig
 from esmerald.template.jinja import JinjaTemplateEngine
 
 template_config = TemplateConfig(