+++ docs_src/middleware/auth_middleware_example.py
@@ -1,4 +1,4 @@
-from esmerald.config.jwt import JWTConfig
+from esmerald.core.config.jwt import JWTConfig
 from esmerald.contrib.auth.saffier.base_user import User
 from esmerald.exceptions import NotAuthorized
 from esmerald.middleware.authentication import AuthResult, BaseAuthMiddleware