+++ tests/app_settings/test_settings_module_string.py
@@ -14,7 +14,7 @@ from esmerald import (
     get,
     settings,
 )
-from esmerald.config import CORSConfig, CSRFConfig
+from esmerald.core.config import CORSConfig, CSRFConfig
 from esmerald.middleware import RequestSettingsMiddleware
 from esmerald.testclient import create_client
 from esmerald.types import Middleware